Prepare project for deploying
